+++ src/fidesops/ops/email_templates/template_names.py
@@ -1,0 +1,1 @@
+SUBJECT_IDENTITY_VERIFICATION_TEMPLATE = "subject_identity_verification.html"